% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.melting.R
\name{print.melting}
\alias{print.melting}
\title{Prints melting temperature from a \code{melting} object}
\usage{
\method{print}{melting}(x, ...)
}
\arguments{
\item{x}{An object of class \code{melting}.}

\item{...}{Unused}
}
\description{
\code{print.melting} prints to console the melting temperature value from an object of
class \code{melting}.
}
\seealso{
\code{\link[rmelting]{melting}}
}
