% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withWE.R
\name{withWE}
\alias{withWE}
\title{Evaluate expression and capture all warnings and errors if any along with results}
\usage{
withWE(expr)
}
\arguments{
\item{expr}{The expression to be evaluated.}
}
\value{
\itemize{
 \item{In cas of Warning(s)}{Returns the value along with the warning message(s).}
 \item{In cas of Error}{Returns NA as the value along with the error message.}
}
}
\description{
Not exported. Strictly internal
}
\examples{
foo <- function(){
  warning("oops")
  1}

foo <- function(){
  warning("oops")
  warning("again oops")
  1}

foo <- function(){
  warning("oops")
  log("a")}

}
\keyword{internal}
