% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withWE.R
\name{withWE}
\alias{withWE}
\title{Evaluate expression and capture all warnings and errors if any along with results}
\usage{
withWE(expr)
}
\arguments{
\item{expr}{The expression to be evaluated.}
}
\description{
Not exported. Strictly internal
}
\keyword{internal}
